# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Periodico, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Detalle_Permisos/ListarDetallePermisosActivosUsuario": {
      "get": {
        "tags": [
          "Detalle_Permisos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Detalle_Permiso"
              }
            }
          }
        }
      }
    },
...
    "/api/Detalle_Permisos/{id}": {
      "get": {
        "tags": [
          "Detalle_Permisos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Detalle_Permiso"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Detalle_Permisos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Detalle_Permisos/Actualizar": {
      "put": {
        "tags": [
          "Detalle_Permisos"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "idpermiso",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "idusuario",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Detalle_Permisos/Crear": {
      "post": {
        "tags": [
          "Detalle_Permisos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AsignarPermisosRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Detalle_Permisos/ActualizarPermisos": {
      "post": {
        "tags": [
          "Detalle_Permisos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdatePermissionsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Permisos/{id}": {
      "get": {
        "tags": [
          "Permisos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Permiso"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Permisos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Permisos/Actualizar": {
      "put": {
        "tags": [
          "Permisos"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "descripcion",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Permisos/Crear": {
      "post": {
        "tags": [
          "Permisos"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "descripcion",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AsignarPermisosRequest": {
      "type": "object",
      "properties": {
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        },
        "permisos": {
          "type": "array",
          "items": {
            "format": "int32",
            "type": "integer"
          }
        }
      }
    },
...
    "Detalle_Permiso": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "estado": {
          "type": "string"
        },
        "idusuario": {
          "format": "int32",
          "type": "integer"
        },
        "idpermiso": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "Permiso": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "descripcion": {
          "type": "string"
        },
        "estado": {
          "type": "string"
        },
        "detallePermisos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Detalle_Permiso"
          }
        }
      }
    },
...
    "UpdatePermissionsRequest": {
      "type": "object",
      "properties": {
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        },
        "addedPermissions": {
          "type": "array",
          "items": {
            "format": "int32",
            "type": "integer"
          }
        },
        "removedPermissions": {
          "type": "array",
          "items": {
            "format": "int32",
            "type": "integer"
          }
        }
      }
    },
...
  }
}